fix: address code review findings (closes #328)
