+++ rector.php
@@ -5,6 +5,6 @@ declare(strict_types=1);
 use Rector\Config\RectorConfig;
 
 return RectorConfig::configure()
-    ->withPaths([__DIR__ . '/src', __DIR__ . '/tests',])
+    ->withPaths([__DIR__ . '/src', __DIR__ . '/tests', __DIR__ . '/benchmarks',])
     ->withPhpSets(php82: true)
     ->withPreparedSets(deadCode: true, codeQuality: true, typeDeclarations: true, symfonyCodeQuality: true);
